+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added fields for nexus and shipping code
+
 ## [0.1.2] - 2022-04-07
 
 ### Added